Move reponse formatter to its own module
